% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abridgeddata_documentaion.R
\docType{data}
\name{Beijing_abridged}
\alias{Beijing_abridged}
\title{Beijing car prices from 2010 to 2015}
\format{
a tibble with 217155 rows and 5 variables
\describe{
\item{year}{year, as an integer between 2010 and 2015}
\item{month}{month, as an integer between 1 and 12}
\item{city}{"Beijing"}
\item{MSRP}{manufacturer's suggested retail price, in RMB}
\item{sales}{number of car models sold in that particular year}
}
}
\source{
Daljord et al. (2021), The Black Market for Beijing License Plates
}
\usage{
Beijing_abridged
}
\description{
A dataset containing the manufacturer's suggested retail price (MSRP) of car
models and how many of these models were sold in Beijing between 2010 and 2015.
}
\details{
Due to privacy restrictions, we cannot distribute the original data sets.
This data set is an anonymized version of the original.
}
\keyword{datasets}
