% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingOT.R
\name{build_costmatrix}
\alias{build_costmatrix}
\title{Create Cost Matrix}
\usage{
build_costmatrix(support, bandwidth = 0)
}
\arguments{
\item{support}{A vector of the common support of source
and target distributions}

\item{bandwidth}{A non-negative number to ignore small transfers;
defaults to 0}
}
\value{
a symmetric, square matrix of dimension \code{length(support)}
}
\description{
Create cost matrix with common \code{support} for both source and target
distributions.
}
\details{
The entries of the cost matrix are given by:
\deqn{c_d(x_1, x_0) = 1( | x_1 - x_0 | > d ),}
where \eqn{x_1} and \eqn{x_0} are entries in \code{support}, \eqn{1(...)}
is the indicator function, and \eqn{d} is \code{bandwidth}.

Note that this function has no internal checks.
The user should be careful to ensure that:
\enumerate{
\item \code{bandwidth} is non-negative.
\item \code{support} contains unique values.
}

Instead of using the common support of our source and target
distributions, \code{build_costmatrix2} uses the support of
the source distribution and the support of the target
distribution separately.
}
\examples{
build_costmatrix(c(0, 1, 2), 1)

}
\seealso{
\code{\link{build_costmatrix2}}
}
