% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingOT.R
\name{build_costmatrix2}
\alias{build_costmatrix2}
\title{Create Cost Matrix with Minimal Supports}
\usage{
build_costmatrix2(support_pre, support_post, bandwidth = 0)
}
\arguments{
\item{support_pre}{A vector of the support of the pre-distribution}

\item{support_post}{A vector of the support of the post-distribution}

\item{bandwidth}{A non-negative number to ignore small transfers;
defaults to 0}
}
\value{
a matrix of dimension \code{length(support_pre)} by
'\code{length(support_post)}
}
\description{
Create cost matrix with \code{support_pre} for pre-distribution and
\code{support_post} for post-distribution.
}
\details{
The \eqn{(i, j)} entries of the cost matrix are given by:
\deqn{c_d(x_i, y_j) = 1( | x_i - y_j | > d ),}
where \eqn{x_i} is the ith entry of \code{support_pre},
\eqn{y_j} is the jth entry of \code{support_post},
\eqn{1(...)} is the indicator function, and
\eqn{d} is \code{bandwidth}.

Note that this function has no internal checks.
The user should be careful to ensure that:
\enumerate{
\item \code{bandwidth} is non-negative.
\item \code{support_pre} contains unique values.
\item \code{support_post} contains unique values.
}

Instead of using the common support of our pre- and
post-distributions as in \code{build_costmatrix}, this function
uses the support of the pre- and post-distributions separately.
Resultantly, this function need not return a square, symmetric
matrix whose diagonals are zero.
}
\examples{
build_costmatrix2(c(0, 1, 2), c(1, 2, 3), 1)

}
\seealso{
\code{\link{build_costmatrix}}
}
