% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fcns.R
\name{check_support}
\alias{check_support}
\title{Check Common Support}
\usage{
check_support(pre_df, post_df, var)
}
\arguments{
\item{pre_df}{A two-column \code{data.frame} describing the pre-distribution;}

\item{post_df}{A two-column \code{data.frame} describing the
post-distribution}

\item{var}{The title of the column of support values in \code{pre_df}
and \code{post_df}}
}
\value{
A named list of three items:
\enumerate{
\item \code{status}: if 1, then \code{pre_df} and \code{post_df}
have invalid/different/repeated entries in \code{var} column
\item \code{message}: describes why \code{status} is 1 or 0
\item \code{support}: if \code{status} is 0, then this is the
common support of \code{pre_df} and \code{post_df}
}
}
\description{
Given two discrete distributions for \code{var}, determine whether
\code{var} has the same unique values.
}
\details{
The distributions \code{pre_df} and \code{post_df} should be of
the following form:
\enumerate{
\item A column called \code{var} for the support of \code{var}
\item A column for the mass for each value of \code{var}
}
Since both distributions are for the same variable, the column
\code{var} must contain the same values.
This function essentially verifies that this is the case.
}
\examples{
pre_df <- data.frame(support = c(1, 2, 3), count = c(1, 2, 3))
post_df <- data.frame(support = c(1, 2, 4), count = c(1, 2, 3))
check_support(pre_df, post_df, support)
}
