% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diftrans.R
\name{diftrans}
\alias{diftrans}
\title{Compute Before-and-After Estimator and Differences-in-Transports Estimator}
\usage{
diftrans(
  pre_main = NULL,
  post_main = NULL,
  pre_control = NULL,
  post_control = NULL,
  var = MSRP,
  count = count,
  bandwidth_vec = seq(0, 40000, 1000),
  minimum_bandwidth = 0,
  maximum_bandwidth = Inf,
  estimator = ifelse(!is.null(pre_control) & !is.null(post_control), "dit", "tc"),
  sims_bandwidth_selection = 0,
  precision = 5e-04,
  sensitivity_lag = 5,
  sensitivity_lead = 5,
  sensitivity_accept = 5,
  sims_subsampling = 0,
  pre_main_subsample_size = NULL,
  post_main_subsample_size = NULL,
  pre_control_subsample_size = NULL,
  post_control_subsample_size = NULL,
  seed = 1,
  conservative = FALSE,
  quietly = FALSE
)
}
\arguments{
\item{pre_main}{probability mass function (see "Details") for \code{var} of the
treated group before treatment occurs}

\item{post_main}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs}

\item{pre_control}{probability mass function (see "Details") for \code{var} of the
control group before treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{post_control}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{var}{the title of the first column of \code{pre_main}, \code{post_main},
\code{pre_control}, and \code{post_control}; default is \code{MSRP}
(see Daljord et al. (2021))}

\item{count}{the title of the second column of \code{pre_main}, \code{post_main},
\code{pre_control}, and \code{post_control}; default is \code{count}
(see Daljord et al. (2021))}

\item{bandwidth_vec}{a vector of bandwidth values to try; default is \code{seq(0, 40000, 1000)}}

\item{minimum_bandwidth}{minimum bandwidth to consider for the estimator;
defaults to 0}

\item{maximum_bandwidth}{maximum bandwidth to consider for the estimator;
defaults to infinity}

\item{estimator}{a string that takes on the value of "dit" for
differences-in-transports estimator or "tc" for the transport cost;
if \code{pre_control} and \code{post_control} are specified, default is "dit";
otherwise, default is "tc"}

\item{sims_bandwidth_selection}{number of simulations for the bandwidth
selection process; defaults to 0}

\item{precision}{threshold to choose the bandwidth; defaults to 0.0005, i.e.,
5 percent.}

\item{sensitivity_lag, sensitivity_lead, sensitivity_accept}{acceptable bandwidths are bandwidths such that among the
\code{sensitivity_lag} previous bandwidths and \code{sensitivity_lead}
bandwidths, \code{sensitivity_accept} of them have a placebo cost that is less
than \code{precision}; TODO: move to Details}

\item{sims_subsampling}{number of subsampling simulations}

\item{pre_main_subsample_size, post_main_subsample_size, pre_control_subsample_size, post_control_subsample_size}{sample size of subsample distributions}

\item{seed}{for reproducibility}

\item{conservative}{if \code{TRUE}, then the bandwidth sequence will be
multiplied by 2 to provide a conservative estimate of the
difference-in-transports estimator; default is \code{FALSE}, only valid
for difference-in-transports estimator}

\item{quietly}{if \code{TRUE}, some results and will be suppressed from printing; default is \code{FALSE}}
}
\value{
a data.frame with the transport costs associated with each value of \code{bandwidth_vec}.
\itemize{
\item \code{bandwidth}: same as \code{bandwidth_vec}
\item \code{main}: transport costs associated with main distributions
\item \code{main2d}: transport costs associated with main distributions using twice the bandwidth;
appears only if \code{conservative = TRUE}
\item \code{control}: transport costs associated with the control distributions;
appears only if \code{pre_control} and \code{post_control}
are specified
\item \code{diff}: \code{main - control}
\item \code{diff2d}: \code{main2d - control}
}
}
\description{
This function implements the before-and-after estimation procedure and the
differences-in-transports estimator described in Daljord et al. (2021).
}
\details{
The pre- and post-distributions given by \code{pre_main}
and \code{post_main} need to be \code{data.frame} objects of
two columns. The first column, whose title should be given by
\code{var}, contains the common support of the two distributions.
In other words, the values in \code{var} in each \code{pre_df}
and \code{post_df} need to be unique and the same across the two
distributions.
The second column named \code{count} provides the mass on the
respective value of the support.
If the sum of \code{count} in each \code{pre_df} and \code{post_df} is 1,
then \code{pre_df} and \code{post_df} are probability mass functions.

The other pair of pre- and post-distributions given by \code{pre_control}
and \code{post_control} follow the same conventions.
In particular, their column names must also be \code{var} and \code{count}.

If \code{pre_main} and \code{post_main} are specified but \code{pre_control}
and \code{post_control} are not, then the function computes the
before-and-after estimator between \code{pre_main} and \code{post_main}.
If \code{pre_control} and \code{post_control} are also specified, then
the function computes the differences-in-transports estimator.

When \code{conservative} is set to TRUE, the differences-in-transports
estimator uses twice the bandwidth when computing the optimal transport cost
between the main distributions.

Each entry of \code{bandwidth_vec} should be a non-negative number.
If mass is transferred less than this number, we ignore these transfers.
Otherwise, we place equal weight on the transfers.
As a result, the transport costs to be used by our estimators can be viewed
as the percentage of mass that has been transferred by more than a
a bandwidth's distance away.

TODO: add documentation about bandwidth selection, min and max bw args,
subsampling procedure
}
\examples{
# Find conservative transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 0
# # step 1: find support
support_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Beijing)
# # step 2: prepare probability mass functions
pre_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
# # step 3: compute results
tc <- diftrans(pre_Beijing, post_Beijing, conservative = TRUE, bandwidth = 0)
tc$main2d

# Find transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 10000
# tc_10000 <- diftrans(pre_Beijing, post_Beijing, bandwidth = 10000)# tc_10000$main
# Find conservative differences-in-transport estimator using Tianjin as a control
# # step 1: find support
support_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Tianjin)
# # step 2: prepare probability mass functions
pre_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
# # step 3: compute results
dit <- diftrans(pre_Beijing, post_Beijing, pre_Tianjin, post_Tianjin,
                   conservative = TRUE, bandwidth = seq(0, 40000, 1000),
                   save_result = TRUE)
dit$optimal_bandwidth
dit$dit
}
