% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingOT.R
\name{get_results}
\alias{get_results}
\title{Obtain Transport Costs and Differences-in-Transports Estimator}
\usage{
get_results(
  pre_main = NULL,
  post_main = NULL,
  pre_control = NULL,
  post_control = NULL,
  var = MSRP,
  bandwidth_seq = seq(0, 40000, 1000),
  estimator = ifelse(!is.null(pre_control) & !is.null(post_control), "dit", "tc"),
  conservative = F,
  quietly = F,
  suppress_progress_bar = F
)
}
\arguments{
\item{pre_main}{probability mass function (see "Details") for \code{var} of the
treated group before treatment occurs}

\item{post_main}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs}

\item{pre_control}{probability mass function (see "Details") for \code{var} of the
control group before treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{post_control}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{var}{the title of the first column of \code{pre_main}, \code{post_main},
\code{pre_control}, and \code{post_control}; default is \code{MSRP}
(see Daljord et al. (2020))}

\item{bandwidth_seq}{a vector of bandwidth values to try; default is \code{seq(0, 40000, 1000)}}

\item{estimator}{a string that takes on the value of "dit" for
differences-in-transports estimator or "tc" for the transport cost;
if \code{pre_control} and \code{post_control} are specified, default is "dit";
otherwise, default is "tc"}

\item{conservative}{if \code{TRUE}, then the bandwidth sequence will be
multiplied by 2 to provide a conservative estimate of the transport costs/
difference-in-transports estimator; default is \code{FALSE}}

\item{quietly}{if \code{TRUE}, some results and will be suppressed from printing; default is \code{FALSE}}

\item{suppress_progress_bar}{if \code{TRUE}, the progress bar will be suppressed; default is \code{FALSE}}
}
\value{
a data.frame with the transport costs associated with each value of \code{bandwidth_seq}.
\itemize{
\item \code{bandwidth}: same as \code{bandwidth_seq}
\item \code{main}: transport costs associated with main distributions
\item \code{main2d}: transport costs associated with main distributions using twice the bandwidth;
appears only if \code{conservative = TRUE}
\item \code{control}: transport costs associated with the control distributions;
appears only if \code{pre_control} and \code{post_control} are specified
\item \code{diff}: \code{main - control}
\item \code{diff2d}: \code{main2d - control}
}
}
\description{
Given the pre and post probability mass functions as well as a vector of
bandwidths, this function returns the associated transport costs.
If another set of pre and post probability mass functions are given for the
control group, then the differences-in-transports estimator is returned.
}
\details{
The \code{pre_main}, \code{post_main}, \code{pre_control}, and
\code{post_control} variables are all probability mass functions.
That is, they are a tibble with two columns:
\itemize{
\item column 1 contains the support of \code{var}, and
\item column 2, which should be titled "counts", contains the corresponding
mass of each value in the support.
}
}
\examples{
# Find conservative transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 0
support_Beijing <- prep_data(Beijing_cleaned, "support",
                     lowerdate = "2010-01-01", upperdate = "2012-01-01")
pre_Beijing <- prep_data(Beijing_cleaned, "pmf", support = support_Beijing,
                         lowerdate = "2010-01-01", upperdate = "2011-01-01")
post_Beijing <- prep_data(Beijing_cleaned, "pmf", support = support_Beijing,
                          lowerdate = "2011-01-01", upperdate = "2012-01-01")
tc <- get_results(pre_Beijing, post_Beijing, conservative = TRUE, bandwidth = 0)
tc$main2d

# Find transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 10000
tc_10000 <- get_results(pre_Beijing, post_Beijing, bandwidth = 10000)
tc_10000$main
# Note: this result is the same as the before-and-after estimator in Daljord et al. (2020)

# Find conservative differences-in-transport estimator using Tianjin as a control
support_Tianjin <- prep_data(Tianjin_cleaned, "support",
                     lowerdate = "2010-01-01", upperdate = "2012-01-01")
pre_Tianjin <- prep_data(Tianjin_cleaned, "pmf", support = support_Tianjin,
                         lowerdate = "2010-01-01", upperdate = "2011-01-01")
post_Tianjin <- prep_data(Tianjin_cleaned, "pmf", support = support_Tianjin,
                          lowerdate = "2011-01-01", upperdate = "2012-01-01")
dit <- get_results(pre_Beijing, post_Beijing, pre_Tianjin, post_Tianjin,
                   conservative = TRUE, bandwidth = seq(0, 40000, 1000))
# Note: the message that is printed contains the differences-in-tranports estimator
# that is in Daljord et al. (2020).


}
