% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diftrans.R
\name{is.diftrans}
\alias{is.diftrans}
\title{Check if object is of class "diftrans"}
\usage{
is.diftrans(x)
}
\arguments{
\item{x}{An object}
}
\value{
TRUE if \code{x} is a "diftrans" object; FALSE otherwise
}
\description{
Check if object is of class "diftrans"
}
