% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fcns.R
\name{preliminaries}
\alias{preliminaries}
\title{Preliminary Error Checks}
\usage{
preliminaries(
  pre_main,
  post_main,
  pre_control,
  post_control,
  var,
  count,
  bandwidth_vec,
  minimum_bandwidth,
  maximum_bandwidth,
  estimator,
  sims_bandwidth_selection,
  precision,
  sensitivity_lag,
  sensitivity_lead,
  sensitivity_accept,
  sims_subsampling,
  pre_main_subsample_size,
  post_main_subsample_size,
  pre_control_subsample_size,
  post_control_subsample_size,
  seed,
  conservative,
  quietly,
  show_progress
)
}
\arguments{
\item{pre_main}{A two-column \code{data.frame} describing the
pre-distribution of the treated observations}

\item{post_main}{A two-column \code{data.frame} describing the
post-distribution of the treated observations}

\item{pre_control}{A two-column \code{data.frame} describing the
pre-distribution of the untreated observations; only required for the
differences-in-transports estimator}

\item{post_control}{A two-column \code{data.frame} describing the
post-distribution of the untreated observations; only required for the
differences in transports estimator}

\item{var}{The title of the common support columns of \code{pre_main},
\code{post_main}, \code{pre_control}, and \code{post_control};
defaults to \code{MSRP}}

\item{count}{The title of the second column of \code{pre_main},
\code{post_main}, \code{pre_control}, and \code{post_control};
defaults to \code{count}}

\item{bandwidth_vec}{A vector of non-negative bandwidth values;
defaults to \code{seq(0, 40000, 1000)}}

\item{minimum_bandwidth}{Minimum bandwidth to consider for the estimator;
defaults to 0}

\item{maximum_bandwidth}{Maximum bandwidth to consider for the estimator;
defaults to infinity}

\item{estimator}{A string that takes on the value of "dit" for
differences-in-transports estimator or "tc" for the transport cost;
if \code{pre_control} and \code{post_control} are specified,
default is "dit";
otherwise, default is "tc"}

\item{sims_bandwidth_selection}{Number of simulations for the bandwidth
selection process; defaults to 0}

\item{precision}{Threshold to choose the bandwidth; defaults to 0.0005, i.e.,
5 percent.}

\item{sensitivity_lag}{See "Details" to understand how these parameters influence bandwidth
selection}

\item{sensitivity_lead}{See "Details" to understand how these parameters influence bandwidth
selection}

\item{sensitivity_accept}{See "Details" to understand how these parameters influence bandwidth
selection}

\item{sims_subsampling}{Number of subsampling simulations}

\item{pre_main_subsample_size}{Size of subsample distributions}

\item{post_main_subsample_size}{Size of subsample distributions}

\item{pre_control_subsample_size}{Size of subsample distributions}

\item{post_control_subsample_size}{Size of subsample distributions}

\item{seed}{Seed for reproducibility; see \code{\link{set.seed}}}

\item{conservative}{If TRUE, then the bandwidth sequence will be
multiplied by 2 to provide a conservative estimate of the
difference-in-transports estimator; default is FALSE, only valid
for difference-in-transports estimator}

\item{quietly}{If TRUE, some messages will be suppressed from printing;
defaults FALSE}

\item{show_progress}{If \code{TRUE}, placebo and subsampling simulations
will show progress; defaults to FALSE}
}
\value{
Either a list of error messages, or a list of modified objects:
\enumerate{
\item estimator
\item conservative
\item count columns from distributions
\item totals from distributions
\item subsample sizes
}
}
\description{
Check arguments of \code{\link{diftrans}} and modify them if necessary
}
\details{
This function uses \code{\link{collect_msg}} to store error messages so that
users get a list of potential problems to fix rather than debugging errors
one at a time.
}
\seealso{
\code{\link{collect_msg}}
}
