% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fcns.R
\name{send_note}
\alias{send_note}
\title{Send note to console}
\usage{
send_note(note, quietly, fcn)
}
\arguments{
\item{note}{A string that will be sent to the console}

\item{quietly}{If FALSE, message will be sent to console}

\item{fcn}{A function, e.g., \code{\link[base]{message}}
or \code{\link[base]{warning}}}
}
\value{
A note in the console if \code{quietly} is FALSE
}
\description{
Send a message or warning if \code{quietly} is FALSE
}
\details{
The intended values of \code{fcn} are \code{\link[base]{message}} and
\code{\link[base]{warning}}.
If \code{quietly} is FALSE, then send the message to the console according
to \code{fcn}. Otherwise, do not send anything to the console.
}
\examples{
\dontrun{
send_note("This is a warning.", FALSE, warning)
}

}
\seealso{
Other debugging: 
\code{\link{collect_msg}()}
}
\concept{debugging}
