% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{Shijiazhuang_sample}
\alias{Shijiazhuang_sample}
\title{Shijiazhuang car prices from 2010 to 2015}
\format{
a tibble with 194468 rows and 16 variables
\describe{
\item{id}{a number that uniquely identifies each Shijiazhuang car model by year}
\item{year}{year, as an integer between 2010 and 2015}
\item{month}{month, as an integer between 1 and 12}
\item{ym}{A Date-class variable of the format yyyy-mm-01}
\item{city}{"Shijiazhuang"}
\item{MSRP}{manufacturer's suggested retail price, in RMB}
\item{sales}{number of car models sold in that particular year}
\item{swtprice}{price, in RMB 1000}
\item{color, noticenum}{uniquely identify car model}
\item{postBeijing}{1 if cars were sold after Beijing lottery in 2011, and 0 otherwise}
\item{postTianjin}{1 if cars were sold after Tianjin lottery/auction in 2014, and 0 otherwise}
\item{Beijing}{1 if cars were sold in Beijing, which is no car in this dataset}
\item{Tianjin}{1 if cars were sold in Tianjin, which is no car in this dataset}
\item{Shijiazhuang}{1 if cars were sold in Shijiazhuang, which is every car in this dataset}
\item{policy_dummy}{0 because there is no policy enacted in Shijiazhuang between 2010 and 2015}
}
}
\source{
Daljord et al. (2020), The Black Market for Tianjin License Plates
}
\usage{
Shijiazhuang_sample
}
\description{
A dataset containing the manufacturer's suggested retail price (MSRP) of car
models and how many of these models were sold in Shijiazhuang between 2010 and 2015.
}
\details{
Due to privacy restrictions, we cannot distribute the original data sets.
This data set is an anonymized version of the original.
}
\keyword{datasets}
