% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fcns.R
\name{valid_bandwidths}
\alias{valid_bandwidths}
\title{importFrom dplyr lag
importFrom dplyr lead}
\usage{
valid_bandwidths(vec, lag, lead, accept, precision)
}
\description{
importFrom dplyr lag
importFrom dplyr lead
}
