% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computingOT.R
\name{build_costmatrix}
\alias{build_costmatrix}
\title{Create Cost Matrix}
\usage{
build_costmatrix(support, bandwidth = 0)
}
\arguments{
\item{support}{A vector of the common support of pre-
and post-distributions}

\item{bandwidth}{A non-negative number to ignore small transfers;
defaults to 0}
}
\value{
a symmetric, square matrix of dimension \code{length(support)}
}
\description{
Create cost matrix with common \code{support} for both pre- and
post-distributions.
}
\details{
The \eqn{(i, j)} entries of the cost matrix are given by:
\deqn{c_d(x_i, x_j) = 1( | x_i - x_j | > d ),}
where \eqn{x_i} and \eqn{x_j} are the i-th and j-th entries in
\code{support}, \eqn{1(...)} is the indicator function, and
\eqn{d} is \code{bandwidth}.

Note that this function has no internal checks.
The user should be careful to ensure that:
\enumerate{
\item \code{bandwidth} is non-negative.
\item \code{support} contains unique values.
}

Instead of using the common support of our pre- and
post-distributions, \code{build_costmatrix2} uses the support of
the pre-distribution and the support of the
post-distribution separately.

Since this function uses the common support, the diagonals
of the cost matrix will refer to the cost of no movement.
For any choice of the \code{bandwidth}, the diagonals
will therefore be zero.
Another consequence of using the common support is that
cost matrix will be symmetric and square.
}
\examples{
\dontrun{
build_costmatrix(c(0, 1, 2), 1)
}

}
\seealso{
\code{\link{build_costmatrix2}}
}
