% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diftrans.R
\name{diftrans}
\alias{diftrans}
\title{Compute Before-and-After Estimator and Differences-in-Transports Estimator}
\usage{
diftrans(
  pre_main = NULL,
  post_main = NULL,
  pre_control = NULL,
  post_control = NULL,
  var = MSRP,
  count = count,
  bandwidth_vec = seq(0, 40000, 1000),
  minimum_bandwidth = 0,
  maximum_bandwidth = Inf,
  estimator = ifelse(!is.null(pre_control) & !is.null(post_control), "dit", "tc"),
  sims_bandwidth_selection = 0,
  precision = 5e-04,
  sensitivity_lag = 5,
  sensitivity_lead = 5,
  sensitivity_accept = 5,
  sims_subsampling = 0,
  pre_main_subsample_size = NULL,
  post_main_subsample_size = NULL,
  pre_control_subsample_size = NULL,
  post_control_subsample_size = NULL,
  seed = 1,
  conservative = FALSE,
  quietly = TRUE,
  show_progress = FALSE
)
}
\arguments{
\item{pre_main}{A two-column \code{data.frame} describing the
pre-distribution of the treated observations}

\item{post_main}{A two-column \code{data.frame} describing the
post-distribution of the treated observations}

\item{pre_control}{A two-column \code{data.frame} describing the
pre-distribution of the untreated observations; only required for the
differences-in-transports estimator}

\item{post_control}{A two-column \code{data.frame} describing the
post-distribution of the untreated observations; only required for the
differences in transports estimator}

\item{var}{The title of the common support columns of \code{pre_main},
\code{post_main}, \code{pre_control}, and \code{post_control};
defaults to \code{MSRP}}

\item{count}{The title of the second column of \code{pre_main},
\code{post_main}, \code{pre_control}, and \code{post_control};
defaults to \code{count}}

\item{bandwidth_vec}{A vector of non-negative bandwidth values;
defaults to \code{seq(0, 40000, 1000)}}

\item{minimum_bandwidth}{Minimum bandwidth to consider for the estimator;
defaults to 0}

\item{maximum_bandwidth}{Maximum bandwidth to consider for the estimator;
defaults to infinity}

\item{estimator}{A string that takes on the value of "dit" for
differences-in-transports estimator or "tc" for the transport cost;
if \code{pre_control} and \code{post_control} are specified,
default is "dit";
otherwise, default is "tc"}

\item{sims_bandwidth_selection}{Number of simulations for the bandwidth
selection process; defaults to 0}

\item{precision}{Threshold to choose the bandwidth; defaults to 0.0005, i.e.,
5 percent.}

\item{sensitivity_lag, sensitivity_lead, sensitivity_accept}{See "Details" to understand how these parameters influence bandwidth
selection}

\item{sims_subsampling}{Number of subsampling simulations}

\item{pre_main_subsample_size, post_main_subsample_size, pre_control_subsample_size, post_control_subsample_size}{Size of subsample distributions}

\item{seed}{Seed for reproducibility; see \code{\link{set.seed}}}

\item{conservative}{If TRUE, then the bandwidth sequence will be
multiplied by 2 to provide a conservative estimate of the
difference-in-transports estimator; default is FALSE, only valid
for difference-in-transports estimator}

\item{quietly}{If TRUE, some messages will be suppressed from printing;
defaults FALSE}

\item{show_progress}{If \code{TRUE}, placebo and subsampling simulations
will show progress; defaults to FALSE}
}
\value{
an object of the "diftrans" class with the following information:
\itemize{
\item \code{bandwidth_vec}: sorted vector of unique bandwidth values used
for evaluating ground cost
\item \code{est}: either "ba" for the before-and-after estimator
or "dit" for the differences-in-transports estimator
\item \code{pre_main_total}: total count in \code{pre_main}
\item \code{post_main_total}: total count in \code{post_main}
\item \code{pre_control_total}: total count in \code{pre_control}
\item \code{post_control_total}: total count in \code{post_control}
\item \code{sims_bandwidth_selection}: number of simulations to run for
bandwidth selection
\item \code{pre_main_subsample_size},
\code{post_main_subsample_size},
\code{post_control_subsample_size},
\code{post_control_subsample_size}: size of subsample distributions
\item \code{conservative}: TRUE or FALSE, depending on whether we use
twice the bandwidth for the treated distributions in the
differences-tranpsorts-estimator
\item \code{seed}: seed (see \code{\link{set.seed}})
\item \code{main_support}: common support of the treated distributions
\item \code{control_support}: common support of the untreated distributions
\item \code{empirical_table}: data frame of transport costs at each
bandwidth in \code{bandwidth_vec} as well as indicators for the
candidate bandwidths and the final bandwidth chosen for the estimate
(\code{optimal_bandwidth})
\item \code{candidate_bandwidths}: vector of candidate bandwidths that
survive the bandwidth selection
\item \code{optimal_bandwidth}: final bandwidth for estimator
\item \code{empirical_cost}: value of estimator at \code{optimal_bandwidth}
\item \code{subsample}: vector of costs that arise from subsampling
\item \code{call}: function call to produce the object of class "diftrans"
}
}
\description{
This function implements the before-and-after estimation procedure and the
differences-in-transports estimator described in Daljord et al. (2021).
}
\details{
The pre- and post-distributions given by \code{pre_main}
and \code{post_main} need to be \code{data.frame} objects of
two columns. The first column, whose title should be given by
\code{var}, contains the common support of the two distributions.
In other words, the values in \code{var} in each \code{pre_df}
and \code{post_df} need to be unique and the same across the two
distributions.
The second column named \code{count} provides the mass on the
respective value of the support.
If the sum of \code{count} in each \code{pre_df} and \code{post_df} is 1,
then \code{pre_df} and \code{post_df} are probability mass functions.

The other pair of pre- and post-distributions given by \code{pre_control}
and \code{post_control} follow the same conventions.
In particular, their column names must also be \code{var} and \code{count}.

If \code{pre_main} and \code{post_main} are specified but \code{pre_control}
and \code{post_control} are not, then the function computes the
before-and-after estimator between \code{pre_main} and \code{post_main}.
If \code{pre_control} and \code{post_control} are also specified, then
the function computes the differences-in-transports estimator.

When \code{conservative} is set to TRUE, the differences-in-transports
estimator uses twice the bandwidth when computing the optimal transport cost
between the main distributions.

Each entry of \code{bandwidth_vec} should be a non-negative number.
If mass is transferred less than this number, we ignore these transfers.
Otherwise, we place equal weight on the transfers.
As a result, the transport costs to be used by our estimators can be viewed
as the percentage of mass that has been transferred by more than a
a bandwidth's distance away.

The bandwidth selection process chooses a bandwidth among
\code{bandwidth_vec} to filter out sampling variation.
The selection process differs by the estimator, but for both estimators,
placebo costs are computed for each bandwidth in every simulation.
We identify bandwidths whose mean placebo costs across all simulations
is less than \code{precision}.
For each such bandwidth, we ask whether \code{sensitivity_accept} of the
following \code{sensitivity_lead} and preceding \code{sensitivity_lag}
bandwidths in \code{bandwidth_vec} are also less than \code{precision}.
If this is the case, then we consider such a bandwidth as a candidate
for the estimator.
Note that we also require candidate bandwidths to be at least
\code{minimum_bandwidth} and no more than \code{maximum_bandwidth}.

For the before-and-after estimator, we choose the smallest of the candidate
bandwidths.
For the differences-in-transports estimator, we choose the candidate
bandwidth with the largest value of the estimator.

For the subsampling procedure, we sample \code{pre_main_subsample_size}
observations from the pre-distribution of the treated units
without replacement.
Similarly, we sample \code{post_main_subsample_size} observations from
the post-distribution without replacement
If we are computing the differences-in-transports estimator, we do the
the same with the utreated analogues.
}
\examples{
# Compute transport cost between Beijing 2010 and Beijing 2011
# with bandwidth = 0
## step 1: find support
support_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Beijing)
## step 2: prepare distributions
pre_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
## step 3: compute results
tc <- diftrans(pre_Beijing, post_Beijing, conservative = TRUE, bandwidth = 0)
tc$empirical_cost

# Find conservative differences-in-transport estimator using Tianjin
# as a control
## step 1: find support
support_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Tianjin)
## step 2: prepare distributions
pre_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
## step 3: compute results
dit <- diftrans(pre_Beijing, post_Beijing, pre_Tianjin, post_Tianjin,
                   conservative = TRUE, bandwidth = seq(0, 40000, 1000))
dit$optimal_bandwidth
dit$empirical_cost
}
