% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diftrans.R
\name{diftrans}
\alias{diftrans}
\title{Obtain Transport Costs and Differences-in-Transports Estimator}
\usage{
diftrans(
  pre_main = NULL,
  post_main = NULL,
  pre_control = NULL,
  post_control = NULL,
  var = MSRP,
  count = count,
  bandwidth_seq = seq(0, 40000, 1000),
  minimum_bandwidth = 0,
  maximum_bandwidth = Inf,
  estimator = ifelse(!is.null(pre_control) & !is.null(post_control), "dit", "tc"),
  sims_bandwidth_selection = 0,
  precision = 5e-04,
  sensitivity_lag = 5,
  sensitivity_lead = 5,
  sensitivity_accept = 5,
  sims_subsampling = 0,
  subsample_pre_main_size = NULL,
  subsample_post_main_size = NULL,
  subsample_pre_control_size = NULL,
  subsample_post_control_size = NULL,
  seed = 1,
  conservative = FALSE,
  quietly = FALSE,
  suppress_progress_bar = FALSE,
  save_result = FALSE,
  costm_main = NULL,
  costm_ref_main = NULL,
  costm_control = NULL,
  costm_ref_control = NULL
)
}
\arguments{
\item{pre_main}{probability mass function (see "Details") for \code{var} of the
treated group before treatment occurs}

\item{post_main}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs}

\item{pre_control}{probability mass function (see "Details") for \code{var} of the
control group before treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{post_control}{probability mass function (see "Details") for \code{var} of the
treated group after treatment occurs; only required for the computing the
differences-in-transports estimator}

\item{var}{the title of the first column of \code{pre_main}, \code{post_main},
\code{pre_control}, and \code{post_control}; default is \code{MSRP}
(see Daljord et al. (2021))}

\item{count}{the title of the second column of \code{pre_main}, \code{post_main},
\code{pre_control}, and \code{post_control}; default is \code{count}
(see Daljord et al. (2021))}

\item{bandwidth_seq}{a vector of bandwidth values to try; default is \code{seq(0, 40000, 1000)}}

\item{minimum_bandwidth}{minimum bandwidth to consider for the estimator;
defaults to 0}

\item{maximum_bandwidth}{maximum bandwidth to consider for the estimator;
defaults to infinity}

\item{estimator}{a string that takes on the value of "dit" for
differences-in-transports estimator or "tc" for the transport cost;
if \code{pre_control} and \code{post_control} are specified, default is "dit";
otherwise, default is "tc"}

\item{sims_bandwidth_selection}{number of simulations for the bandwidth
selection process; defaults to 0}

\item{precision}{threshold to choose the bandwidth; defaults to 0.0005, i.e.,
5 percent.}

\item{sensitivity_lag, sensitivity_lead, sensitivity_accept}{acceptable bandwidths are bandwidths such that among the
\code{sensitivity_lag} previous bandwidths and \code{sensitivity_lead}
bandwidths, \code{sensitivity_accept} of them have a placebo cost that is less
than \code{precision}; TODO: move to Details}

\item{sims_subsampling}{number of subsampling simulations}

\item{subsample_pre_main_size, subsample_post_main_size, subsample_pre_control_size, subsample_post_control_size}{sample size of subsample distributions}

\item{seed}{for reproducibility}

\item{conservative}{if \code{TRUE}, then the bandwidth sequence will be
multiplied by 2 to provide a conservative estimate of the
difference-in-transports estimator; default is \code{FALSE}, only valid
for difference-in-transports estimator}

\item{quietly}{if \code{TRUE}, some results and will be suppressed from printing; default is \code{FALSE}}

\item{suppress_progress_bar}{if \code{TRUE}, the progress bar will be suppressed; default is \code{FALSE}}

\item{save_result}{if \code{TRUE}, the estimator as
well as the associated bandwidth will be returned}

\item{costm_main}{if \code{NULL}, the cost matrix with common support will be such that if the transport
distance is greater than what is specified in \code{bandwidth_seq}, cost is 1 and 0 otherwise.}

\item{costm_ref_main}{if \code{NULL}, the cost matrix referenced by \code{transport::transport} will be
using the minimal support of main distributions}

\item{costm_control}{if \code{NULL}, the cost matrix with common support will be such that if the transport
distance is greater than what is specified in \code{bandwidth_seq}, cost is 1 and 0 otherwise.}

\item{costm_ref_control}{if \code{NULL}, the cost matrix referenced by \code{transport::transport} will be
using the minimal support of control distributions}
}
\value{
a data.frame with the transport costs associated with each value of \code{bandwidth_seq}.
\itemize{
\item \code{bandwidth}: same as \code{bandwidth_seq}
\item \code{main}: transport costs associated with main distributions
\item \code{main2d}: transport costs associated with main distributions using twice the bandwidth;
appears only if \code{conservative = TRUE}
\item \code{control}: transport costs associated with the control distributions;
appears only if \code{pre_control} and \code{post_control}
are specified
\item \code{diff}: \code{main - control}
\item \code{diff2d}: \code{main2d - control}
}

If \code{save_result = TRUE}, then a list is returned, with the first element
(labeled \code{out}) being the data.frame described above.
The second element (labeled \code{dit}) is the differences-in-transports
estimator, and the third and final element (labeled \code{optimal_bandwidth})
is the bandwidth associated with the estimator.
}
\description{
Given the pre and post probability mass functions as well as a vector of
bandwidths, this function returns the associated transport costs.
If another set of pre and post probability mass functions are given for the
control group, then the differences-in-transports estimator is returned.
}
\details{
The \code{pre_main}, \code{post_main}, \code{pre_control}, and
\code{post_control} variables are all probability mass functions.
That is, they are a tibble with two columns:
\itemize{
\item column 1 contains the full support of \code{var}, and
\item column 2 contains the corresponding mass of each value
in the support.
}
Since column 1 contains the full support of \code{var} and all these
distributions are of \code{var}, column 1 must be the same for all
distributions.

The cost matrices specified by \code{costm} should use a common support of
the respective distributions.
However, \code{costm_ref} matrices should use the minimal support of the
respective pre and post distributions.
}
\examples{
# Find conservative transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 0
# # step 1: find support
support_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Beijing)
# # step 2: prepare probability mass functions
pre_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Beijing <- Beijing_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
# # step 3: compute results
tc <- diftrans(pre_Beijing, post_Beijing, conservative = TRUE, bandwidth = 0)
tc$main2d

# Find transport cost of MSRP in Beijing between 2010 and 2011 using bandwidth = 10000
# tc_10000 <- diftrans(pre_Beijing, post_Beijing, bandwidth = 10000)# tc_10000$main
# Find conservative differences-in-transport estimator using Tianjin as a control
# # step 1: find support
support_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2012-01-01") \%>\%
  dplyr::select(MSRP) \%>\%
  dplyr::distinct() \%>\%
  dplyr::arrange(MSRP) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  unlist()
temp <- data.frame(MSRP = support_Tianjin)
# # step 2: prepare probability mass functions
pre_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2010-01-01") & ym < "2011-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
post_Tianjin <- Tianjin_sample \%>\%
  dplyr::filter(ym >= as.Date("2011-01-01") & ym < "2012-01-01") \%>\%
  dplyr::group_by(dplyr::across(c(MSRP))) \%>\%
  dplyr::summarise(count = sum(sales)) \%>\%
  dplyr::filter(!is.na(MSRP)) \%>\%
  dplyr::left_join(temp, .) \%>\%
  dplyr::select(MSRP, count) \%>\%
  tidyr::replace_na(list(count = 0)) \%>\%
  tibble::as_tibble()
# # step 3: compute results
dit <- diftrans(pre_Beijing, post_Beijing, pre_Tianjin, post_Tianjin,
                   conservative = TRUE, bandwidth = seq(0, 40000, 1000),
                   save_result = TRUE)
dit$optimal_bandwidth
dit$dit
}
